+++ mkosi/distributions/centos_epel.py
@@ -1,7 +1,0 @@
-# SPDX-License-Identifier: LGPL-2.1+
-
-from mkosi.distributions.centos import CentosInstaller
-
-
-class CentosEpelInstaller(CentosInstaller):
-    pass