+++ mkosi/distributions/rocky_epel.py
@@ -1,7 +1,0 @@
-# SPDX-License-Identifier: LGPL-2.1+
-
-from mkosi.distributions.rocky import RockyInstaller
-
-
-class RockyEpelInstaller(RockyInstaller):
-    pass